+++ packages/cli/src/commands/lint/lint.test.ts
@@ -17,7 +17,7 @@ const exampleMessages: Message[] = [
 		variants: [
 			{
 				languageTag: "en",
-				match: {},
+				match: [],
 				pattern: [
 					{
 						type: "Text",
@@ -33,7 +33,7 @@ const exampleMessages: Message[] = [
 		variants: [
 			{
 				languageTag: "en",
-				match: {},
+				match: [],
 				pattern: [
 					{
 						type: "Text",
@@ -43,7 +43,7 @@ const exampleMessages: Message[] = [
 			},
 			{
 				languageTag: "de",
-				match: {},
+				match: [],
 				pattern: [
 					{
 						type: "Text",
@@ -65,7 +65,7 @@ async function setupProject(enabledLintRule?: MessageLintRule) {
 			languageTags: ["en", "de", "it"],
 			modules: ["_mockPlugin.js", ...(enabledLintRule ? ["lintRule.js"] : [])],
 			messageLintRuleLevels: {},
-		} satisfies ProjectSettings),
+		} satisfies ProjectSettings)
 	)
 
 	const _mockPlugin: Plugin = {