+++ packages/rpc/src/functions/machineTranslateMessage.ts
@@ -19,15 +19,15 @@ Promise<Result<Message, string>> {
 				!args.sourceLanguageTag ||
 				!args.message.variants.some((variant) => variant.languageTag === args.sourceLanguageTag)
 			) {
-				throw new Error("Source Language configuration missing")
+				throw new Error("Source language configuration missing")
 			}
 			for (const variant of args.message.variants.filter(
-				(variant) => variant.languageTag === args.sourceLanguageTag,
+				(variant) => variant.languageTag === args.sourceLanguageTag
 			)) {
 				const targetVariant = getVariant(args.message, {
 					where: {
 						languageTag: targetLanguageTag,
-						selectors: variant.match,
+						match: variant.match,
 					},
 				})
 				if (targetVariant) {
@@ -44,7 +44,7 @@ Promise<Result<Message, string>> {
 							format: "html",
 							key: privateEnv.GOOGLE_TRANSLATE_API_KEY,
 						}),
-					{ method: "POST" },
+					{ method: "POST" }
 				)
 				if (!response.ok) {
 					return { error: response.statusText }
@@ -84,7 +84,7 @@ const escapeEnd = "</span>"
 
 function serializePattern(
 	pattern: Message["variants"][number]["pattern"],
-	placeholderMetadata: PlaceholderMetadata,
+	placeholderMetadata: PlaceholderMetadata
 ) {
 	let result = ""
 	for (const [i, element] of pattern.entries()) {