+++ packages/versioned-interfaces/marketplace-manifest/src/interface.ts
@@ -11,27 +11,32 @@ const MarketplaceManifestBase = Type.Object({
 	$schema: Type.Optional(Type.Literal("https://inlang.com/schema/marketplace-manifest")),
 	icon: Type.Optional(Type.String()),
 	coverImage: Type.Optional(
-		Type.String({ description: "The cover displayed in the marketplace." }),
+		Type.String({ description: "The cover displayed in the marketplace." })
+	),
+	gallery: Type.Optional(
+		Type.Array(
+			Type.String({
+				description: "The images displayed in the marketplace. The first image is the cover.",
+			})
+		)
 	),
 	displayName: Translatable(
-		Type.String({ description: "The name which is displayed in the marketplace." }),
+		Type.String({ description: "The name which is displayed in the marketplace." })
 	),
 	description: Translatable(
 		Type.String({
 			description: "The description which is displayed in the marketplace.",
-		}),
+		})
 	),
 	publisherName: Type.String(),
 	publisherIcon: Type.Optional(Type.String()),
 	readme: Translatable(
-		Type.TemplateLiteral("${string}.md", {
-			description: "The link must be a valid markdown file.",
-		}),
+		Type.TemplateLiteral("${string}.md", { description: "The path to the readme file." })
 	),
 	keywords: Type.Array(Type.String()),
 	license: Type.Literal("Apache-2.0"),
 	website: Type.Optional(
-		Type.String({ description: "An optional link to the website of the marketplace item." }),
+		Type.String({ description: "An optional link to the website of the marketplace item." })
 	),
 })