+++ packages/website/src/pages/documentation/EditButton.tsx
@@ -6,7 +6,7 @@ interface EditButtonProps {
 
 export function EditButton(props: EditButtonProps) {
 	return (
-		<div class="md:pl-9 w-full pt-8 sm:pt-12 max-sm:pb-8 flex justify-between">
+		<div class="w-full pt-8 sm:pt-12 max-sm:pb-8 flex justify-between">
 			<a
 				href={props.href}
 				class="text-info/80 hover:text-info/100 text-sm font-semibold flex items-center"