+++ packages/website/src/renderer/state.ts
@@ -13,5 +13,5 @@ import type { PageContext, PageContextRenderer } from "./types.js"
 // eslint-disable-next-line solid/reactivity
 export const [currentPageContext, setCurrentPageContext] = createStore({}) as [
 	PageContext,
-	SetStoreFunction<PageContextRenderer>,
+	SetStoreFunction<PageContextRenderer>
 ]